Fixed bug in calculating X and O's win total
-In Events.js GameHistory, it was showing odd behavior when calculating
how many wins X's and O's had when only 1 open game in database. I
removed if statement that only allowed it to run if more than 1 game
present in database and added another conditional function to that
checks to see if variable are not equal to 0 and if they are it
can count the game.